requirements.txt
flask==2.0.2
pyyaml==5.4.1
python-dotenv==0.19.2